增加多項特徵辨識PuzzleRecognizerImpl
