+++ ExclusiveProgram/ExclusiveProgram.csproj
@@ -129,6 +129,8 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="puzzle.visual\concrete\DefaultPuzzleFactory.cs" />
+    <Compile Include="puzzle.visual\concrete\SiftFlannPuzzleRecognizerImpl.cs" />
+    <Compile Include="puzzle.visual\concrete\SiftBfmPuzzleRecognizerImpl.cs" />
     <Compile Include="puzzle.visual\concrete\PuzzleCorrector.cs" />
     <Compile Include="puzzle.visual\concrete\PuzzleLocator.cs" />
     <Compile Include="puzzle.visual\concrete\PuzzleRecognizer.cs" />
@@ -170,6 +172,7 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
+    <Compile Include="puzzle.visual\concrete\KazeFlannPuzzleRecognizerImpl.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
